+++ projects/framework-flame/source/ftrack_framework_flame/utils/presets.py
